TRINIDAD-16: Resolve issues reported by FindBugs - Use the latest version of FindBugs to analyze the Trinidad codebase and fix reported problems.   Most were stylistic, but a few genuine bugs were detected.  Changes included: - Consistently use valueOf() instead of new Integer/Character/Short/Long/etc. - Avoid unnecessary toString() and new String() calls - A few synchronization issues, none especially significant - Make inner classes static were possible - Make constants static were this was missing - Make clones of arrays (mostly) consistently - Avoid String instance equality - Fix some incorrect null pointer checks - Fix some object equals() implementations

+++ trinidad-impl/src/main/java/org/apache/myfaces/trinidadinternal/renderkit/core/xhtml/ChooseDateRenderer.java
@@ -1546,7 +1546,7 @@ public class ChooseDateRenderer extends XhtmlRenderer
   private static final int _HALF_MAX_CHOICE_ITEMS = _MAX_CHOICE_ITEMS / 2;
 
 
-  private static final int _MILLIS_IN_DAY = 1000 * 60 * 60 * 24;
+  private static final long _MILLIS_IN_DAY = 1000L * 60 * 60 * 24;
 
   private static final long _MIN_TIME;
